fix windows build issue & pass instanceId to test for metric dimension
